#20 refactor: order, company enum core migration , customException으로 변경, feignclient 중복 수정
- enum migration
- customException 추가
- feign client 중복 수정
