+++ company/src/main/java/com/two_ddang/logistics/company/application/service/company/CompanyService.java
@@ -1,5 +1,10 @@
 package com.two_ddang.logistics.company.application.service.company;
 
+import com.two_ddang.logistics.company.application.dtos.company.CompanyDetailResponse;
+import com.two_ddang.logistics.company.application.dtos.company.CompanyResponse;
+import com.two_ddang.logistics.company.application.dtos.company.CreateCompanyResponse;
+import com.two_ddang.logistics.company.application.dtos.company.UpdateCompanyInfoResponse;
+import com.two_ddang.logistics.company.application.exception.BusinessException;
 import com.two_ddang.logistics.company.application.service.product.ProductService;
 import com.two_ddang.logistics.company.domain.model.company.Company;
 import com.two_ddang.logistics.company.domain.model.product.Product;
@@ -10,11 +15,6 @@ import com.two_ddang.logistics.company.infrastrucuture.dtos.RestockHubRequest;
 import com.two_ddang.logistics.company.presentation.dtos.company.CreatedCompanyRequest;
 import com.two_ddang.logistics.company.presentation.dtos.company.RestockRequest;
 import com.two_ddang.logistics.company.presentation.dtos.company.UpdateCompanyInfoRequest;
-import com.two_ddang.logistics.company.application.dtos.company.CompanyDetailResponse;
-import com.two_ddang.logistics.company.application.dtos.company.CompanyResponse;
-import com.two_ddang.logistics.company.application.dtos.company.CreateCompanyResponse;
-import com.two_ddang.logistics.company.application.dtos.company.UpdateCompanyInfoResponse;
-import com.two_ddang.logistics.core.exception.ApplicationException;
 import com.two_ddang.logistics.core.exception.ErrorCode;
 import com.two_ddang.logistics.core.util.ResponseDTO;
 import lombok.RequiredArgsConstructor;
@@ -50,7 +50,7 @@ public class CompanyService {
 
     public CompanyDetailResponse getCompany(UUID companyId) {
         Company company = companyRepository.findById(companyId)
-                .orElseThrow(() -> new ApplicationException(ErrorCode.NOT_FOUND));
+                .orElseThrow(() -> new BusinessException(ErrorCode.NOT_FOUND));
 
         String hubName = hubService.getHubInfo(company.getHubId()).getData().getName();
         return CompanyDetailResponse.of(company, hubName);
@@ -59,7 +59,7 @@ public class CompanyService {
     @Transactional
     public UpdateCompanyInfoResponse updateCompanyInfo(UUID companyId, UpdateCompanyInfoRequest updateCompanyRequestDto) {
         Company company = companyRepository.findById(companyId)
-                .orElseThrow(() -> new ApplicationException(ErrorCode.NOT_FOUND));
+                .orElseThrow(() -> new BusinessException(ErrorCode.NOT_FOUND));
         company.updateInfo(updateCompanyRequestDto);
 
         String hubName = hubService.getHubInfo(company.getHubId()).getData().getName();
@@ -69,7 +69,7 @@ public class CompanyService {
     @Transactional
     public void deleteCompany(UUID companyId) {
         Company company = companyRepository.findById(companyId)
-                .orElseThrow(() -> new ApplicationException(ErrorCode.NOT_FOUND));
+                .orElseThrow(() -> new BusinessException(ErrorCode.NOT_FOUND));
 //        company.delete();
     }
 
@@ -77,9 +77,9 @@ public class CompanyService {
     public void restock(UUID companyId, RestockRequest restockRequest) {
         // 재입고 요청 오면 허브 물품 입고 api 호출하게 개발
         Company company = companyRepository.findById(companyId)
-                .orElseThrow(() -> new ApplicationException(ErrorCode.NOT_FOUND));
+                .orElseThrow(() -> new BusinessException(ErrorCode.NOT_FOUND));
         Product product = productRepository.findById(restockRequest.getProductId())
-                .orElseThrow(() -> new ApplicationException(ErrorCode.NOT_FOUND));
+                .orElseThrow(() -> new BusinessException(ErrorCode.NOT_FOUND));
 
         RestockHubRequest request = RestockHubRequest.create(product, restockRequest);
