+++ order/src/main/java/com/two_ddang/logistics/order/infrastructure/CompanyFeignClient.java
@@ -2,10 +2,12 @@ package com.two_ddang.logistics.order.infrastructure;
 
 import com.two_ddang.logistics.core.util.ResponseDTO;
 import com.two_ddang.logistics.order.infrastructure.dtos.CompanyDetailResponse;
+import com.two_ddang.logistics.order.infrastructure.dtos.CompanyProductResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 
+import java.util.List;
 import java.util.UUID;
 
 @FeignClient(name = "company")
@@ -13,4 +15,8 @@ public interface CompanyFeignClient extends CompanyService{
 
     @GetMapping("/api/v1/companies/{companyId}")
     ResponseDTO<CompanyDetailResponse> getCompany(@PathVariable UUID companyId);
+
+    @Override
+    @GetMapping("/api/v1/products/companies/{companyId}")
+    ResponseDTO<List<CompanyProductResponse>> getCompanyProducts(@PathVariable UUID companyId);
 }
