--- order/src/main/java/com/two_ddang/logistics/order/infrastructure/dtos/CompanyType.java
@@ -1,6 +1,0 @@
-package com.two_ddang.logistics.order.infrastructure.dtos;
-
-public enum CompanyType {
-
-    PRODUCER, RECEIVER
-}
