--- order/src/main/java/com/two_ddang/logistics/order/infrastructure/dtos/DeliveryStatus.java
@@ -1,12 +1,0 @@
-package com.two_ddang.logistics.order.infrastructure.dtos;
-
-import lombok.Getter;
-
-@Getter
-public enum DeliveryStatus {
-
-    WAITING,
-    DELIVERING,
-    DELIVERED,
-
-}
