+++ order/src/main/java/com/two_ddang/logistics/order/presentation/dtos/UpdateOrderStatusRequest.java
@@ -1,6 +1,6 @@
 package com.two_ddang.logistics.order.presentation.dtos;
 
-import com.two_ddang.logistics.order.domain.model.OrderStatus;
+import com.two_ddang.logistics.core.entity.OrderStatus;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
